+++ src/main/kotlin/no/nav/sosialhjelp/innsyn/saksoversikt/SaksDetaljerResponse.kt
@@ -1,0 +1,8 @@
+package no.nav.sosialhjelp.innsyn.saksoversikt
+
+data class SaksDetaljerResponse(
+    val fiksDigisosId: String,
+    val soknadTittel: String,
+    val status: String,
+    val antallNyeOppgaver: Int?
+)
