+++ src/main/kotlin/no/nav/sosialhjelp/innsyn/saksoversikt/SaksListeResponse.kt
@@ -1,4 +1,4 @@
-package no.nav.sosialhjelp.innsyn.domain
+package no.nav.sosialhjelp.innsyn.saksoversikt
 
 import com.fasterxml.jackson.annotation.JsonFormat
 import java.util.Date
@@ -9,11 +9,4 @@ data class SaksListeResponse(
     @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS")
     val sistOppdatert: Date,
     val kilde: String
-)
-
-data class SaksDetaljerResponse(
-    val fiksDigisosId: String,
-    val soknadTittel: String,
-    val status: String,
-    val antallNyeOppgaver: Int?
 )
