+++ src/main/kotlin/no/nav/sosialhjelp/innsyn/saksoversikt/SaksOversiktController.kt
@@ -1,21 +1,13 @@
-package no.nav.sosialhjelp.innsyn.rest
+package no.nav.sosialhjelp.innsyn.saksoversikt
 
 import no.finn.unleash.Unleash
 import no.nav.security.token.support.core.api.ProtectedWithClaims
-import no.nav.sosialhjelp.api.fiks.DigisosSak
 import no.nav.sosialhjelp.api.fiks.exceptions.FiksException
-import no.nav.sosialhjelp.innsyn.app.featuretoggle.DIALOG_UNDERSOKELSE_GRUPPE_1
-import no.nav.sosialhjelp.innsyn.app.featuretoggle.DIALOG_UNDERSOKELSE_GRUPPE_2
-import no.nav.sosialhjelp.innsyn.app.featuretoggle.DIALOG_UNDERSOKELSE_GRUPPE_3
-import no.nav.sosialhjelp.innsyn.app.featuretoggle.DIALOG_UNDERSOKELSE_GRUPPE_4
 import no.nav.sosialhjelp.innsyn.app.featuretoggle.FAGSYSTEM_MED_INNSYN_I_PAPIRSOKNADER
-import no.nav.sosialhjelp.innsyn.app.subjecthandler.SubjectHandlerUtils
 import no.nav.sosialhjelp.innsyn.digisosapi.FiksClient
 import no.nav.sosialhjelp.innsyn.digisossak.oppgaver.OppgaveService
 import no.nav.sosialhjelp.innsyn.digisossak.saksstatus.SaksStatusService
 import no.nav.sosialhjelp.innsyn.domain.InternalDigisosSoker
-import no.nav.sosialhjelp.innsyn.domain.SaksDetaljerResponse
-import no.nav.sosialhjelp.innsyn.domain.SaksListeResponse
 import no.nav.sosialhjelp.innsyn.domain.SaksStatus
 import no.nav.sosialhjelp.innsyn.domain.SoknadsStatus
 import no.nav.sosialhjelp.innsyn.event.EventService
@@ -57,10 +49,10 @@ class SaksOversiktController(
         val responselist = saker
             .map {
                 SaksListeResponse(
-                    it.fiksDigisosId,
-                    "Søknad om økonomisk sosialhjelp",
-                    unixTimestampToDate(it.sistEndret),
-                    IntegrationUtils.KILDE_INNSYN_API
+                    fiksDigisosId = it.fiksDigisosId,
+                    soknadTittel = "Søknad om økonomisk sosialhjelp",
+                    sistOppdatert = unixTimestampToDate(it.sistEndret),
+                    kilde = IntegrationUtils.KILDE_INNSYN_API
                 )
             }
             .sortedByDescending { it.sistOppdatert }
@@ -78,37 +70,6 @@ class SaksOversiktController(
         }
 
         return ResponseEntity.ok().body(responselist)
-    }
-
-    @GetMapping("/sisteSak")
-    fun hentSisteSak(@RequestHeader(value = HttpHeaders.AUTHORIZATION) token: String): ResponseEntity<DigisosSak> {
-        tilgangskontroll.sjekkTilgang(token)
-
-        val saker = try {
-            fiksClient.hentAlleDigisosSaker(token)
-        } catch (e: FiksException) {
-            return ResponseEntity.status(503).build()
-        }
-
-        val sisteSoknad = saker.sortedByDescending { it.originalSoknadNAV?.timestampSendt }.firstOrNull()
-            ?: return ResponseEntity.noContent().build()
-
-        val lastDigit: Char = SubjectHandlerUtils.getUserIdFromToken()[10]
-
-        if (unleashClient.isEnabled(DIALOG_UNDERSOKELSE_GRUPPE_1) && listOf('0', '1', '2').contains(lastDigit)) {
-            return ResponseEntity.ok().body(sisteSoknad)
-        }
-        if (unleashClient.isEnabled(DIALOG_UNDERSOKELSE_GRUPPE_2) && listOf('3', '4').contains(lastDigit)) {
-            return ResponseEntity.ok().body(sisteSoknad)
-        }
-        if (unleashClient.isEnabled(DIALOG_UNDERSOKELSE_GRUPPE_3) && listOf('5', '6', '7').contains(lastDigit)) {
-            return ResponseEntity.ok().body(sisteSoknad)
-        }
-        if (unleashClient.isEnabled(DIALOG_UNDERSOKELSE_GRUPPE_4) && listOf('8', '9').contains(lastDigit)) {
-            return ResponseEntity.ok().body(sisteSoknad)
-        }
-
-        return ResponseEntity.noContent().build()
     }
 
     @GetMapping("/saksDetaljer")
